Updated api route to save tickets and first message

+++ tickets_app/database/migrations/2020_02_03_11_create_direccions_table.php
@@ -19,7 +19,7 @@ class CreateDireccionsTable extends Migration
             $table->smallInteger('numExt');
             $table->smallInteger('numInt')->nullable();
             $table->string('colonia', 50);
-            $table->smallInteger('cp');
+            $table->integer('cp');
             $table->unsignedBigInteger('idEstado');
             $table->unsignedBigInteger('idMunicipio');
